Update code layout
